Add files via upload
The VoltageOverTime.xlsx is the original file from Mack, and I transformed the file to format of csv as vot.csv. 
The voltageOverTime.R is the script file I used to analyze the data. The models I tried are included in the R script file, and I included my comments about the data and models.